.claude/skills/bdd-scaffold/test-suite/pyproject.toml
[project]
name = "bdd-test-suite"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "behave>=1.2.6",
    "databricks-sdk>=0.40.0",
    "httpx>=0.27.0",
]

[tool.behave]
default_format = "pretty"
show_timings = true
default_tags = ["not @wip and not @skip"]
logging_level = "INFO"

[tool.behave.userdata]
warehouse_id = "auto"
catalog = "main"
